requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.10.1
python-multipart==0.0.6

# Deep Learning & AI
tensorflow==2.15.0
transformers==4.36.0
huggingface-hub==0.19.4

# Computer Vision & Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
pydicom==2.4.3
scikit-image==0.22.0

# Data Processing
numpy==1.24.3
pandas==2.1.4

# Web & API
python-dotenv==1.0.0
aiofiles==23.2.1

# Utilities
requests==2.31.0
tqdm==4.66.1
loguru==0.7.2 